Refactor QueryReferenceTests and add VersionedQueryReferenceTests for improved coverage and clarity
